refactor : provider -> utils 전환
