+++ src/main/java/roomescape/auth/JwtConfig.java
@@ -1,0 +1,12 @@
+package auth;
+
+import auth.JwtUtils;
+
+@Configuration
+public class JwtConfig {
+
+    @Bean
+    public JwtUtils jwtUtils(@Value("${jwt.secret}") String secretKey) {
+        return new JwtUtils(secretKey);
+    }
+}
